+++ types/test.ts
@@ -156,7 +156,13 @@ const transaction = new StellarSdk.TransactionBuilder(account, {
   .setMinAccountSequence("5")
   .setMinAccountSequenceAge(5)
   .setMinAccountSequenceLedgerGap(5)
-  .setExtraSigners([sourceKey.publicKey()])
+  .setExtraSigners([
+    sourceKey.publicKey(),
+    StellarSdk.SignerKey.encodeSignedPayloadFromAddress(
+      sourceKey.publicKey(),
+      Buffer.from("hello, world!", "ascii")
+    )
+  ])
   .build(); // $ExpectType () => Transaction<Memo<MemoType>, Operation[]>
 
 const transactionFromXDR = new StellarSdk.Transaction(transaction.toEnvelope(), StellarSdk.Networks.TESTNET); // $ExpectType Transaction<Memo<MemoType>, Operation[]>